refactor: move related modules to their folders
