+++ src/entities/course/index.ts
@@ -1,4 +1,4 @@
-export type { ApiCoursesIds, Course, CourseItemData, CourseStatus, Section } from './types';
+export type { ApiCoursesIds, Course, CourseItemData, CourseStatus } from './types';
 export { CourseCard } from './ui/course-card/course-card';
 export { CourseItem } from './ui/course-item/course-item';
 export { courseStore } from './model/store';
