+++ src/widgets/about-course/index.ts
@@ -1,3 +1,4 @@
+export type { AboutCourseSectionData } from './types';
 export { AboutCourseSection } from '@/widgets/about-course/ui/about-course-section/about-course-section';
 export { isAboutCourseSection } from './helpers/is-about-course-section';
 export { transformAboutCourseSection } from './helpers/transform-about-course-section';
