+++ src/widgets/about-course/ui/about-course-section/about-course-section.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames/bind';
 
-import { AboutCourseSectionData } from '@/entities/course/types';
 import { LinkCustom } from '@/shared/ui/link-custom';
 import { Paragraph } from '@/shared/ui/paragraph';
 import { WidgetTitle } from '@/shared/ui/widget-title';
+import { AboutCourseSectionData } from '@/widgets/about-course';
 import { GridItem } from '@/widgets/about-course/ui/grid-item/grid-item';
 
 import styles from './about-course-section.module.scss';
