+++ src/widgets/learning-path-stages/index.ts
@@ -1,3 +1,4 @@
+export type { LearningPathStagesSectionData } from './types';
 export { LearningPathStageItem } from './ui/learning-path-stage-item/learning-path-stage-item';
 export { LearningPathStages } from './ui/learning-path-stages/learning-path-stages';
 export { isLearningPathStagesSection } from './helpers/is-learning-path-stages-section';
