+++ src/widgets/media-text-block/index.ts
@@ -1,3 +1,4 @@
+export type { MediaTextBlockSectionData } from './types';
 export { MediaTextBlock } from './ui/media-text-block';
 export { isMediaTextBlockSection } from './helpers/is-media-text-block-section';
 export { transformMediaTextBlockSection } from './helpers/transform-media-text-block-section';
