+++ src/widgets/media-text-block/types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from 'react';
+
+export type MediaTextBlockSectionData = {
+  title: string;
+  contentLeft?: ReactNode;
+  contentRight?: ReactNode;
+  linkUrl?: string | null;
+  linkLabel?: string;
+  disabledLinkLabel?: string;
+  backgroundColor?: string;
+};
