+++ src/widgets/video-block/index.ts
@@ -1,3 +1,4 @@
+export type { VideoBlockSectionData } from './types';
 export { VideoBlock } from './ui/video-block';
 export { isVideoBlockSection } from './helpers/is-video-block-section';
 export { transformVideoBlockSection } from './helpers/transform-video-block-section';
