+++ src/widgets/video-block/types.ts
@@ -1,0 +1,5 @@
+export type VideoBlockSectionData = {
+  title: string;
+  url: string;
+  videoTitle: string;
+};
